+++ mopidy/core/actor.py
@@ -7,6 +7,7 @@ import pykka
 
 from mopidy import audio, backend, mixer
 from mopidy.audio import PlaybackState
+from mopidy.core.history import HistoryController
 from mopidy.core.library import LibraryController
 from mopidy.core.listener import CoreListener
 from mopidy.core.playback import PlaybackController
@@ -22,6 +23,10 @@ class Core(
     library = None
     """The library controller. An instance of
     :class:`mopidy.core.LibraryController`."""
+
+    history = None
+    """The playback history controller. An instance of
+    :class:`mopidy.core.HistoryController`."""
 
     playback = None
     """The playback controller. An instance of
@@ -41,6 +46,8 @@ class Core(
         self.backends = Backends(backends)
 
         self.library = LibraryController(backends=self.backends, core=self)
+
+        self.history = HistoryController()
 
         self.playback = PlaybackController(
             mixer=mixer, backends=self.backends, core=self)